fix(react): components exporting default and named exports
