+++ packages/react/src/components/Accordion/AccordionItem.tsx
@@ -1,14 +1,8 @@
 import classNames from "classnames";
-import { createContext, FC } from "react";
+import { FC } from "react";
 import useGlobalSettings from "../../hooks/useGlobalSettings";
-import {
-  AccordionItemContextProps,
-  AccordionItemProps,
-} from "./AccordionItem.props";
-
-export const AccordionItemContext = createContext(
-  {} as AccordionItemContextProps
-);
+import { AccordionItemContext } from "./AccordionCtx";
+import { AccordionItemProps } from "./AccordionItem.props";
 
 const AccordionItem: FC<AccordionItemProps> = ({
   children,
