+++ packages/react/src/components/GlobalProvider/GlobalCtx.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { GlobalProviderContextProps } from "./GlobalProvider.props";
+
+export const GlobalProviderContext = createContext(
+  {} as GlobalProviderContextProps
+);
