+++ packages/react/src/components/GlobalProvider/GlobalProvider.tsx
@@ -1,13 +1,7 @@
-import { createContext, FC } from "react";
 import { themeprefix } from "@ilo-org/themes/tokens/theme/base.json";
-import {
-  GlobalProviderProps,
-  GlobalProviderContextProps,
-} from "./GlobalProvider.props";
-
-export const GlobalProviderContext = createContext(
-  {} as GlobalProviderContextProps
-);
+import { FC } from "react";
+import { GlobalProviderContext } from "./GlobalCtx";
+import { GlobalProviderProps } from "./GlobalProvider.props";
 
 const GlobalProvider: FC<GlobalProviderProps> = ({
   prefix = themeprefix.value,
