+++ packages/react/src/components/GlobalProvider/index.ts
@@ -1,4 +1,2 @@
-export {
-  default as GlobalProvider,
-  GlobalProviderContext,
-} from "./GlobalProvider";
+export { default as GlobalProvider } from "./GlobalProvider";
+export { GlobalProviderContext } from "./GlobalCtx";
