+++ packages/react/src/components/List/ListCtx.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+import { ListContextProps } from "./List.props";
+import { ListItemContextProps } from "./ListItem.props";
+
+export const ListContext = createContext({} as ListContextProps);
+
+export const ListItemContext = createContext({} as ListItemContextProps);
