+++ packages/react/src/components/Tag/TagCtx.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { TagSetContextProps } from "./TagSet.props";
+
+export const TagSetContext = createContext({} as TagSetContextProps);
